feat(#9): merge user services and form components
