--- admin-tool/src/ts/modules/users/ts/components/edit-user/edit-user.component.less
@@ -1,75 +1,0 @@
-@import (reference) "../../../../../../css/variables.less";
-@import "../../../../../../css/base-modal.less";
-
-.modal {
-  display: block;
-
-  .modal-dialog {
-    width: 50%;
-    min-width: 30rem;
-    max-width: 50rem;
-    margin: 5% auto;
-  }
-
-  .modal-body {
-    max-height: 28rem;
-    overflow-y: auto;
-
-    .form-group {
-      margin-bottom: 0.9375rem;
-
-      label {
-        font-weight: bold;
-        margin-bottom: 0.3125rem;
-        display: block;
-      }
-
-      &.required label::after {
-        content: ' *';
-        color: @red;
-      }
-
-      &.has-error {
-        .form-control {
-          border-color: @red;
-        }
-        .help-block {
-          color: @red;
-        }
-      }
-
-      .help-block {
-        font-size: 0.75rem;
-        margin-top: 0.25rem;
-        display: block;
-      }
-
-      .list-unstyled {
-        margin: 0;
-
-        .checkbox {
-          margin: 0.3125rem 0;
-
-          label {
-            font-weight: normal;
-            cursor: pointer;
-
-            input[type='checkbox'] {
-              margin-right: 0.375rem;
-            }
-          }
-        }
-      }
-
-      .input-group {
-        .btn-default {
-          border-color: @gray-light;
-
-          &:hover {
-            background-color: @gray-ultra-lighter;
-          }
-        }
-      }
-    }
-  }
-}
